+++ src/driver/eeprom.h
@@ -9,7 +9,7 @@
  * @param[in] addr The address to write to.
  * @param[in] data The byte to write.
  */
-void eeprom_write(uint8_t addr, uint8_t data);
+void eeprom_write(uint16_t addr, uint8_t data);
 
 /**
  * Read a byte of data from the EEPROM.
@@ -17,6 +17,6 @@ void eeprom_write(uint8_t addr, uint8_t data);
  * @param[in] addr The address to read from.
  * @return The byte of data.
  */
-uint8_t eeprom_read(uint8_t addr);
+uint8_t eeprom_read(uint16_t addr);
 
 #endif
